Fixes JQuery loading error
As the JQuery library is loaded at the end of the bpositive.blade.php layout,
the views could not include scripts that use it, as it is loaded after the
scripts.

This commit adds the "endscripts" section in the main layout, that allows the
views to include their own scripts after the JQuery and Bootstrap libraries are
loaded.