Changes the tab order in transcription view
The tab order has been changed in the transcription view, mainly to show the
PSS view as the first element. In addition, Summary has been moved to the
second place, as it shows a general summary of what happened, and the notes
view has been moved to the last place, as it usually does not contain
information.