Disables the lineupNodes option in the TreeView
The lineupNodes option in the PhyD3 cause that all the branch names to appear
aligned at the end of the phylogram, joining each name with the end of each
branch with a dotted line. This option has been disabled, and, now, the branch
names appear at the end of each branch.